requirement.txt
streamlit
openai
python-dotenv
plyer